+++ openai-client/src/commonMain/kotlin/com.aallam.openai.client/internal/api/EditsApi.kt
@@ -5,18 +5,16 @@ import com.aallam.openai.api.edits.EditsRequest
 import com.aallam.openai.client.Edits
 import com.aallam.openai.client.internal.http.HttpRequester
 import com.aallam.openai.client.internal.http.perform
-import io.ktor.client.call.body
-import io.ktor.client.request.post
-import io.ktor.client.request.setBody
-import io.ktor.client.request.url
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
+import io.ktor.client.call.*
+import io.ktor.client.request.*
+import io.ktor.http.*
 
 /**
  * Implementation of [Edits]
  */
 internal class EditsApi(private val requester: HttpRequester) : Edits {
 
+    @Deprecated("Edits is deprecated. Chat completions is the recommend replacement.")
     override suspend fun edit(request: EditsRequest): Edit {
         return requester.perform {
             it.post {