+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/completion/Choice.kt
@@ -1,5 +1,6 @@
 package com.aallam.openai.api.completion
 
+import com.aallam.openai.api.core.FinishReason
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -27,11 +28,12 @@ public data class Choice(
      * The log probabilities of the chosen tokens and the top [CompletionRequest.logprobs] tokens.
      */
     @SerialName("logprobs")
+    @Deprecated("removed from the response")
     public val logprobs: Logprobs? = null,
 
     /**
      * The reason why GPT-3 stopped generating, for example "length".
      */
     @SerialName("finish_reason")
-    public val finishReason: String? = null,
+    public val finishReason: FinishReason,
 )