+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/finetune/HyperParams.kt
@@ -30,4 +30,25 @@ public data class HyperParams(
      * training when completions are short.
      */
     @SerialName("prompt_loss_weight") val promptLossWeight: Double,
+
+
+    /**
+     * If set, we calculate classification-specific metrics such as accuracy and F-1 score using the validation set at
+     * the end of every epoch. These metrics can be viewed in the results file.
+     *
+     * In order to compute classification metrics, you must provide a `validation_file`. Additionally, you must specify
+     * [classificationNClasses] for multiclass classification or [classificationPositiveClass] for binary classification.
+     */
+    @SerialName("compute_classification_metrics") val computeClassificationMetrics: Boolean? = null,
+
+    /**
+     * The number of classes in a classification task.
+     */
+    @SerialName("classification_n_classes") val classificationNClasses: Int? = null,
+
+    /**
+     * The positive class in binary classification.
+     * This parameter is needed to generate precision, recall, and F1 metrics when doing binary classification.
+     */
+    @SerialName("classification_positive_class") val classificationPositiveClass : String? = null,
 )